requirements.txt
google-cloud-aiplatform>=1.25.0
google-cloud-storage>=2.10.0
kfp>=2.0.0
pandas>=1.5.0
pillow>=9.0.0
numpy>=1.23.0
